# go/docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: 65070209@kmitl.ac.th
    name: API Support
    url: https://github.com/Project-DSView/backend
  description: This is the backend API for DSView application In Services GO
  termsOfService: http://swagger.io/terms/
  title: DSView API Go Service
  version: 1.0.0-alpha
paths:
...
  /api/queue/execute:
    post:
      consumes:
      - application/json
      description: Submit code for execution in the queue
      parameters:
      - description: Code execution data
        in: body
        name: execution
        required: true
        schema:
          properties:
            code:
              type: string
            exercise_id:
              type: string
            file_name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Code submitted for execution
          schema:
            properties:
              data:
                properties:
                  job_id:
                    type: string
                type: object
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Submit code for execution
      tags:
      - queue
  /api/queue/jobs:
    get:
      description: Get queue jobs with filtering and pagination (Teachers and TAs can see jobs by course, students see only their own)
      parameters:
      - description: Queue type filter
        enum:
        - code_execution
        - code_review
        in: query
        name: type
        type: string
      - description: Status filter
        enum:
        - pending
        - processing
        - completed
        - failed
        - cancelled
        in: query
        name: status
        type: string
      - description: Course ID filter (Teachers and TAs only)
        in: query
        name: course_id
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Queue jobs retrieved successfully
          schema:
            properties:
              data:
                properties:
                  jobs:
                    items:
                      type: object
                    type: array
                  pagination:
                    type: object
                type: object
              success:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Get queue jobs
      tags:
      - queue
  /api/queue/jobs/{id}:
    get:
      description: Get detailed information about a specific queue job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Queue job retrieved successfully
          schema:
            properties:
              data:
                type: object
              success:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Job not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Get queue job by ID
      tags:
      - queue
  /api/queue/jobs/{id}/cancel:
    post:
      description: Cancel a pending queue job (students can cancel their own, teachers
        can cancel any)
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Job cancelled successfully
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Job not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Cancel a queue job
      tags:
      - queue
  /api/queue/jobs/{id}/process:
    post:
      description: Manually process a queue job (Teachers and TAs only)
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Job processed successfully
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Job not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Process a queue job (Teacher/TA only)
      tags:
      - queue
  /api/queue/review:
    post:
      consumes:
      - application/json
      description: Submit code for review in the queue
      parameters:
      - description: Code review data
        in: body
        name: review
        required: true
        schema:
          properties:
            course_id:
              type: string
            exercise_id:
              type: string
            review_notes:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Code submitted for review
          schema:
            properties:
              data:
                properties:
                  job_id:
                    type: string
                type: object
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Submit code for review
      tags:
      - queue
  /api/queue/stats:
    get:
      description: Get statistics about the queue (Teachers and Admins only)
      produces:
      - application/json
      responses:
        "200":
          description: Queue statistics retrieved successfully
          schema:
            properties:
              data:
                type: object
              success:
                type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get queue statistics
      tags:
      - queue
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication for secure endpoints.
    in: header
    name: dsview-api-key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"